refactor: extract TranscriptionResult.swift from TranscriptionModels.swift
Splits the `TranscriptionResult` value out of TranscriptionModels.swift
so that file holds only TranscriptionRequest + TranscriptionSegment.

Byte-identical move of lines 22-37. Both files import Foundation.

TranscriptionModels.swift: 42 → 26 lines. Tests: 667.
