+++ Sources/BugNarrator/Services/TranscriptionModels.swift
@@ -19,22 +19,6 @@ struct TranscriptionRequest: Sendable {
     }
 }
 
-struct TranscriptionResult: Sendable {
-    let text: String
-    let segments: [TranscriptionSegment]
-    let qualityFindings: [TranscriptQualityFinding]
-
-    init(
-        text: String,
-        segments: [TranscriptionSegment],
-        qualityFindings: [TranscriptQualityFinding] = []
-    ) {
-        self.text = text
-        self.segments = segments
-        self.qualityFindings = qualityFindings
-    }
-}
-
 struct TranscriptionSegment: Decodable, Sendable {
     let start: Double
     let end: Double
